--- .env
@@ -1,1 +1,0 @@
-DB_URL=postgres://localuser:admin@localhost:5432/localdb
